ekzページのデザインのekz欄のコンテナ側を修正

+++ entrypoint/entrypoint-ekz.sh
@@ -42,9 +42,11 @@ killProcess () {
     port=$1
     headerSample="PID"
     pid=$(lsof -i :$port | grep -v $headerSample | head -n 1 | awk '{print $2}')
-    execCmd="kill -9 $pid"
-    echo "$execCmd"
-    bash -c "$execCmd"
+    if [ "$pid" <> "" ]; then
+        execCmd="kill -9 $pid"
+        echo "$execCmd"
+        bash -c "$execCmd"
+    fi
 }
 
 echo "========== start ekz_api(rails) =========="
